Переносит всю логику по отрисовке точек маршрута из render/events.js в controllers/trip-controller.js
