+++ src/controllers/trip-controller.js
@@ -1,4 +1,166 @@
-import {renderDaysWithEvents} from "../render/events";
+import {render, RenderPosition, replace} from "../utils/render.js";
+import {ESCAPE_KEY, ESC_KEY} from "../const.js";
+import DayComponent from "../components/day.js";
+import DayInfoComponent from "../components/day-info.js";
+import EventsComponent from "../components/events.js";
+import EventComponent from "../components/event.js";
+import EventEditComponent from "../components/event-edit.js";
+
+// Логика для формирования дней
+// Формирует массив с начальными датами событий.
+const getEventsStartDates = (events) => {
+  let startDates = [];
+  for (const event of events) {
+    startDates.push(event.time.startTime);
+  }
+  return startDates;
+};
+
+// Возвращает объект с данными о дне, месяце и годe.
+const getFullDate = (date) => {
+  return {
+    date: date.getDate(),
+    month: date.getMonth(),
+    year: date.getFullYear()
+  };
+};
+// Возвращает массив строк с уникальными датами в виде строк.
+const getEventsUniqueDates = (dates) => {
+  const datesWithoutTime = dates.map((date) => getFullDate(date));
+  let uniqueDates = new Set();
+  for (const date of datesWithoutTime) {
+    uniqueDates.add(JSON.stringify(date));
+  }
+  return Array.from(uniqueDates);
+};
+
+// Получает дату в виде строки "{"date": "1", "month": "11", "year": "2020"}"
+// и массив с объектами-событиями.
+// Возвращает массив объектов-событий, соответствующих дате.
+const groupEventsByStartDate = (dateString, eventsArray) => {
+  const dateObj = JSON.parse(dateString);
+  const {date, month, year} = dateObj;
+  const properDate = new Date(year, month, date);
+  const nextDate = new Date(year, month, date + 1);
+
+  return eventsArray.filter((eventItem) => {
+    return eventItem.time.startTime >= properDate && eventItem.time.startTime < nextDate;
+  });
+};
+
+// Для каждого дня (уникальной начальной даты) нужно сформировать разметку с выводом
+// шаблона для дня и шаблона событий, соответствующих дню.
+// Если сортировка не по event, то числа нужно спрятать. Все события будут выводиться в один
+// шаблон дня.
+
+// Events выводятся в days
+// Сформируем структуру данных, в которой будут события по дате, фильтр, начальные даты.
+const prepareDaysWithEventsBeforeRendering = (events, sort) => {
+  sort = document.querySelector(`input[name=trip-sort]:checked`);
+  const daysAndEvents = [];
+  let daysCount = 0;
+
+  const sortedByStartDayEvents = events.slice().sort(
+      function (a, b) {
+        return a.time.startTime - b.time.startTime;
+      }
+  );
+
+  if (sort.value === `sort-event`) {
+    const startDates = getEventsStartDates(sortedByStartDayEvents);
+    const uniqDates = getEventsUniqueDates(startDates);
+
+    for (const uniqDate of uniqDates) {
+      const eventsByDate = groupEventsByStartDate(uniqDate, sortedByStartDayEvents);
+      daysCount++;
+
+      daysAndEvents.push({
+        eventSort: true,
+        daysCount,
+        uniqDate,
+        events: eventsByDate
+      });
+    }
+  } else {
+    daysAndEvents.push({
+      eventSort: false,
+      daysCount: 1,
+      uniqDate: ``,
+      events: sortedByStartDayEvents
+    });
+  }
+
+  return daysAndEvents;
+};
+
+// в EventsComponent отрисовать EventComponent (<li class="trip-events__item"> (events) </li>)
+const renderEvent = (eventsListElement, event) => {
+
+  const replaceEventToEdit = () => {
+    replace(eventEditComponent, eventComponent);
+  };
+
+  const replaceEditToEvent = () => {
+    replace(eventComponent, eventEditComponent);
+  };
+
+  const onEscKeyDown = (evt) => {
+    const isEscKey = evt.key === ESCAPE_KEY || evt.key === ESC_KEY;
+
+    if (isEscKey) {
+      replaceEditToEvent();
+      document.removeEventListener(`keydown`, onEscKeyDown);
+    }
+  };
+
+  const eventComponent = new EventComponent(event);
+  eventComponent.setOpenButtonClickHandler(() => {
+    replaceEventToEdit();
+    document.addEventListener(`keydown`, onEscKeyDown);
+  });
+
+  const eventEditComponent = new EventEditComponent(event);
+  eventEditComponent.setEventEditFormSubmitHandler((evt) => {
+    evt.preventDefault();
+    replaceEditToEvent();
+    document.removeEventListener(`keydown`, onEscKeyDown);
+  });
+
+  render(eventsListElement, eventComponent, RenderPosition.BEFOREEND);
+};
+
+const renderDayInfo = (dayComponent, eventSort, daysCount, uniqDate) => {
+  render(dayComponent.getElement(), new DayInfoComponent(eventSort, daysCount, uniqDate), RenderPosition.BEFOREEND);
+};
+
+const renderEvents = (dayComponent, events) => {
+  render(dayComponent.getElement(), new EventsComponent(events), RenderPosition.BEFOREEND);
+  const eventsListElement = dayComponent.getElement().querySelector(`.trip-events__list`);
+
+  events.forEach((event) => {
+    renderEvent(eventsListElement, event);
+  });
+};
+
+// отрисовать DayComponent (<li class="trip-days__item  day"></li>)
+//   - в DayComponent отрисовать DayInfoComponent (<div class="day__info"> (count, dateString) </div>)
+//   - в DayComponent же отрисовть EventsComponent (<ul class="trip-events__list"></ul>)
+//      - в EventsComponent отрисовать EventComponent (<li class="trip-events__item"> (events) </li>)
+const renderDaysWithEvents = (tripDaysComponent, allEvents) => {
+  const daysWithEvents = prepareDaysWithEventsBeforeRendering(allEvents);
+
+  daysWithEvents.forEach((dayWithEvents) => {
+    const {eventSort, daysCount, uniqDate, events} = dayWithEvents;
+    // отрисовать DayComponent (<li class="trip-days__item day"></li>)
+    const dayComponent = new DayComponent();
+    render(tripDaysComponent.getElement(), dayComponent, RenderPosition.BEFOREEND);
+    //    в DayComponent отрисовать DayInfoComponent (<div class="day__info"> (count, dateString) </div>)
+    renderDayInfo(dayComponent, eventSort, daysCount, uniqDate);
+    //    в DayComponent отрисовть EventsComponent (<ul class="trip-events__list"></ul>)
+    renderEvents(dayComponent, events);
+  });
+};
+
 
 export default class TripController {
   constructor(container) {
